add more ui test cases
