+++ Volumio/Main.storyboard
@@ -445,6 +445,9 @@
                                     <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                                 </view>
                                 <blurEffect style="extraLight"/>
+                                <userDefinedRuntimeAttributes>
+                                    <userDefinedRuntimeAttribute type="string" keyPath="accessibilityIdentifier" value="blur-overlay"/>
+                                </userDefinedRuntimeAttributes>
                             </visualEffectView>
                             <progressView opaque="NO" contentMode="scaleToFill" verticalHuggingPriority="750" progressViewStyle="bar" translatesAutoresizingMaskIntoConstraints="NO" id="VvN-0w-wtO">
                                 <rect key="frame" x="0.0" y="355" width="375" height="2.5"/>
@@ -499,6 +502,9 @@
                                                             <constraint firstAttribute="width" secondItem="AZ4-6W-RsB" secondAttribute="height" multiplier="1:1" id="vTj-5m-EfR"/>
                                                         </constraints>
                                                         <state key="normal" image="play"/>
+                                                        <userDefinedRuntimeAttributes>
+                                                            <userDefinedRuntimeAttribute type="string" keyPath="accessibilityIdentifier" value="play"/>
+                                                        </userDefinedRuntimeAttributes>
                                                         <connections>
                                                             <action selector="pressPlay:" destination="QqN-dO-5fx" eventType="touchUpInside" id="8bG-t4-E5k"/>
                                                         </connections>
